Update the repo to more modern toolchain (#21)
* Update to Espressif8266 2.6.3 toolchain

This necessitates the removal (temporarily, at least) of -Werror so we
can complete a compilation successfully.

* Update mkdata.py for Python 3

Python 2.7 has been sunsetted since January 1, 2020. Update the builder
to use Python 3.9 as it is the latest stable version.

* Use GitHub Actions

Since my balance went negative on Travis-CI, it has never built anything for me.
Switch to GitHub Actions which works arguably better than Travis.

* GitHub Actions: remove staging env

The staging environment does not work, so remove it from CI for now.